backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
boto3==1.34.0
PyPDF2==3.0.1
opensearch-py==2.4.0
python-dotenv==1.0.0
PyMuPDF==1.23.0
camelot-py[cv]==0.11.0